rag-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.13
langchain-community==0.0.29
langchain-text-splitters==0.0.1
langchain-groq==0.1.0
chromadb==0.4.24
sentence-transformers==2.5.1
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
bs4==0.0.2
lxml==5.1.0
